requirements.txt
torch>=2.0
torchvision>=0.15
Pillow>=9.0
matplotlib>=3.5
numpy
pandas
scikit-learn
tqdm
